internal/http-server/handlers/url/remove/remove.go
package remove

import (
	"github.com/go-chi/chi/v5"
	"github.com/go-chi/chi/v5/middleware"
	"github.com/go-chi/render"
	"log/slog"
	"net/http"
	resp "url-shortener/internal/lib/api/response"
)

type Response struct {
	resp.Response
	Alias string `json:"alias,omitempty"`
}

//go:generate go run github.com/vektra/mockery/v2@v2.44.1 --name=UrlRemover
type UrlRemover interface {
	RemoveUrl(alias string) (int64, error)
}

func New(log *slog.Logger, urlRemover UrlRemover) http.HandlerFunc {
	return func(w http.ResponseWriter, r *http.Request) {
		const op = "handlers.url.remove.New"

		log := log.With(
			slog.String("op", op),
			slog.String("request_id", middleware.GetReqID(r.Context())),
		)

		alias := chi.URLParam(r, "alias")
		if alias == " " {
			log.Info("alias is empty")

			render.JSON(w, r, resp.Error("invalid request"))

			return
		}

		urlRes, err := urlRemover.RemoveUrl(alias)
		if err != nil {
			render.JSON(w, r, resp.Error(err.Error()))
		}

		if urlRes == 0 {
			render.JSON(w, r, resp.Error("Url with this alias not found"))

			return
		}

		responseOk(w, r, alias)
	}
}

func responseOk(w http.ResponseWriter, r *http.Request, alias string) {
	render.JSON(w, r, Response{
		Response: resp.Ok(),
		Alias:    alias,
	})
}
